Added some data checking for the input file
FileInput now checks for data in the input file to make sure its the
right kind of data.  Checks each item in the #rotues except the arrival
and departure airports.  If it doesnt pass nothing is loaded from the
input, the route table is empty and a message is posted in the console.
Main file also modified slightly to work with this.