Add Route and Find Route Updated
Adding a route is fully complete now. A new route can be added to the
route table. It just cannot be saved to a text file. It also has error
checking and only one issue. Whenever a "d" is put into the dollar text
box it is recorded for some reason. Also, the Find route is merely just
started. I added the functionality to recognize what check box is being
selected as well as only search for the departure time. I have also
inserted that into the history area. I also edited the Main class so
that the create table can be reusable, since it will need to be used
again for the search panel.