Total Commit
For some reason I was having a lot of issues with my eclipse so I had to
reinstall and do some other stuff to get it working. Here is all of the
documents, including functionality to open the user manual when the user
manual is clicked under the help menu.